+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - [`bootstrap` CHANGELOG](bootstrap/CHANGELOG.md)
 - [`relay` CHANGELOG](relay/CHANGELOG.md)
 - [`client` CHANGELOG](client/CHANGELOG.md)
+- [`crawler` CHANGELOG](crawler/CHANGELOG.md)